chore: make a naming scheme
For widgets with logic, the logic goes under ${capture}_logic.dart.
Right now, however there aren't any logical widgets except history.
And history isn't logical, it just does funny stuff.
The `_key = GlobalKey()` I mean.
And `initState()`.
Whateva.
