+++ lib/widgets/history/history.dart
@@ -1,79 +1,1 @@
-import "package:flutter/material.dart";
-import "package:flutter_riverpod/flutter_riverpod.dart";
-
-import "../../model.dart";
-
-class HistoryListView extends ConsumerStatefulWidget {
-  const HistoryListView({
-    super.key,
-  });
-
-  @override
-  HistoryListViewState createState() => HistoryListViewState();
-}
-
-class HistoryListViewState extends ConsumerState<HistoryListView> {
-  /// Needed so that [AppState] can tell [AnimatedList] below to animate
-  /// new items.
-  final _key = GlobalKey();
-
-  @override
-  void initState() {
-    super.initState();
-    // "ref" can be used in all life-cycles of a StatefulWidget.
-    ref.read(globalAppStateProvider);
-  }
-
-  /// Used to "fade out" the history items at the top, to suggest continuation.
-  static const Gradient _maskingGradient = LinearGradient(
-    // This gradient goes from fully transparent to fully opaque black...
-    colors: [Colors.transparent, Colors.black],
-    // ... from the top (transparent) to half (0.5) of the way to the bottom.
-    stops: [0.0, 0.5],
-    begin: Alignment.topCenter,
-    end: Alignment.bottomCenter,
-  );
-
-  @override
-  Widget build(BuildContext context) {
-    final appState = ref.watch(globalAppStateProvider);
-    Future(() {
-      ref.read(globalAppStateProvider.notifier).setKey(_key);
-    });
-
-    return ShaderMask(
-      shaderCallback: (bounds) => _maskingGradient.createShader(bounds),
-      // This blend mode takes the opacity of the shader (i.e. our gradient)
-      // and applies it to the destination (i.e. our animated list).
-      blendMode: BlendMode.dstIn,
-      child: AnimatedList(
-        key: _key,
-        reverse: true,
-        padding: const EdgeInsets.only(top: 100),
-        initialItemCount: appState.history.length,
-        itemBuilder: (context, index, animation) {
-          final pair = appState.history[index];
-          return SizeTransition(
-            sizeFactor: animation,
-            child: Center(
-              child: TextButton.icon(
-                onPressed: () async {
-                  ref
-                      .read(globalAppStateProvider.notifier)
-                      .toggleFavorite(pair);
-                },
-                icon: appState.favorites.contains(pair)
-                    ? const Icon(Icons.favorite, size: 12)
-                    : const SizedBox(),
-                label: Text(
-                  pair.asLowerCase,
-                  semanticsLabel: pair.asPascalCase,
-                ),
-              ),
-            ),
-          );
-        },
-      ),
-    );
-  }
-}
+export "history_view.dart";
